# API/Rendu API/Nouveau dossier/API/openapi-main.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
  description: This is the API documentation for the application.
servers:
  - url: http://localhost:3001
    description: Local server
paths:
...
  /authors:
    get:
      tags:
      - Authors
        - User
      operationId: listAuthors
      summary: "List all authors"
      parameters:
        - name: "sortBy"
          in: "query"
          description: "Sort by field"
          required: false
          schema:
            type: "string"
        - name: "sortDirection"
          in: "query"
          description: "Sort direction"
          required: false
          schema:
            type: "string"
            enum: ["asc", "desc"]
      responses:
        '200':
          description: A list of authors
    post:
      tags:
      - Authors
        - Admin
      operationId: createAuthor
      summary: "Create a new author"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullname, birthdate]
              properties:
                fullname:
                  type: string
                birthdate:
                  type: string
                deathdate:
                  type: string
      responses:
        '200':
          description: Author created          
  /authors/{id}:
    get:
      tags:
      - Authors
        - User
      operationId: getAuthorById
      summary: "Get author by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Author found
        '404':
          description: Author not found
    delete:
      tags:
      - Authors
        - Admin
      operationId: deleteAuthor
      summary: "Delete author by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Author deleted
        '404':
          description: Author not found
    put:
      tags:
      - Authors
        - Admin
      operationId: updateAuthor
      summary: "Update author by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullname:
                  type: string
                birthdate:
                  type: string
                deathdate:
                  type: string
      responses:
        '200':
          description: Author updated successfully
        '400':
          description: Invalid author data
        '404':
          description: Author not found
  /authors/fullname/{fullname}:
    get:
      tags:
      - Authors
        - User
      operationId: getAuthorByFullname
      summary: "Get author by fullname"
      parameters:
        - name: fullname
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Author found
        '404':
          description: Author not found
...
  /reason:
    get:
      tags:
        - Reason
          - Admin
      operationId: listReasons
      summary: "List all reasons"
      parameters:
        - name: "sortBy"
          in: "query"
          description: "Sort by field"
          required: false
          schema:
            type: "string"
        - name: "sortDirection"
          in: "query"
          description: "Sort direction"
          required: false
          schema:
            type: "string"
            enum: ["asc", "desc"]
      responses:
        '200':
          description: A list of reasons
    post:
      tags:
        - Reason
          - Admin
      operationId: createReason
      summary: "Create a new reason"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label, description]
              properties:
                label:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Reason created
  /reason/{id}:
    get:
      tags:
        - Reason
          - Admin
      operationId: getReasonById
      summary: "Get reason by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Reason found
        '404':
          description: Reason not found
    delete:
      tags:
        - Reason
          - Admin
      operationId: deleteReason
      summary: "Delete reason by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Reason deleted
        '404':
          description: Reason not found
    put:
      tags:
        - Reason
          - Admin
      operationId: updateReason
      summary: "Update reason by ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Reason updated successfully
        '400':
          description: Invalid reason data
        '404':
          description: Reason not found
  /reason/label/{label}:
    get:
      tags:
        - Reason
          - Admin
      operationId: getReasonByLabel
      summary: "Get reason by label"
      parameters:
        - name: label
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reason found
        '404':
          description: Reason not found
...